requirements.txt
Django==1.7.4
Pygments==2.0.2
bpython==0.13.2
django-autofixture==0.9.2
requests==2.5.1
